fix(auth): split portclaim by build tag — syscall.Kill is Unix-only (v0.3.2 release fix)
Goreleaser's windows_amd64 + windows_arm64 targets failed to compile because
syscall.Kill doesn't exist on Windows. v0.3.2 tag was pushed but no release
artifacts were published.

Split internal/auth/portclaim.go into three:
- portclaim.go              — platform-agnostic logic (PortHolder, IsOurDaemon,
                              TakeOverPort, function-variable hooks)
- portclaim_unix.go         — !windows: lsof/ps shell-outs + syscall.Kill
                              wiring (production path for macOS + Linux)
- portclaim_windows.go      — windows: stub init() that wires no-op
                              implementations. Effect: gsuite-mcp auth on
                              Windows falls back to the pre-#158 EADDRINUSE
                              error — no worse than before this fix. Real
                              Windows support would need x/sys/windows
                              GetExtendedTcpTable + OpenProcess/TerminateProcess;
                              filed mentally as future work.

Verified:
- go build ./...                      (darwin host)            ✓
- GOOS=windows GOARCH=amd64 go build  (release target)         ✓
- GOOS=windows GOARCH=arm64 go build  (release target)         ✓
- GOOS=linux GOARCH=amd64   go build  (release target)         ✓
- go test ./internal/auth/            (15 tests, all pass)     ✓
